WIP No validation if email verification is disabled

+++ usercp.php
@@ -969,24 +969,36 @@ if($mybb->input['action'] == "do_email" && $mybb->request_method == "post")
 		{
 			if($mybb->user['usergroup'] != "5" && $mybb->usergroup['cancp'] != 1)
 			{
-				$activationcode = random_str();
-				$now = TIME_NOW;
-				$db->delete_query("awaitingactivation", "uid='".$mybb->user['uid']."'");
-				$newactivation = array(
-					"uid" => $mybb->user['uid'],
-					"dateline" => TIME_NOW,
-					"code" => $activationcode,
-					"type" => "e",
-					"oldgroup" => $mybb->user['usergroup'],
-					"misc" => $db->escape_string($mybb->input['email'])
-				);
-				$db->insert_query("awaitingactivation", $newactivation);
-
-				$username = $mybb->user['username'];
 				$uid = $mybb->user['uid'];
+				$username = $mybb->user['username'];
+
+				if($mybb->settings['regtype'] == "verify")
+				{
+					// Emails require verification
+					$activationcode = random_str();
+					$db->delete_query("awaitingactivation", "uid='".$mybb->user['uid']."'");
+
+					$newactivation = array(
+						"uid" => $mybb->user['uid'],
+						"dateline" => TIME_NOW,
+						"code" => $activationcode,
+						"type" => "e",
+						"oldgroup" => $mybb->user['usergroup'],
+						"misc" => $db->escape_string($mybb->input['email'])
+					);
+
+					$db->insert_query("awaitingactivation", $newactivation);
+	
+					$mail_message = $lang->sprintf($lang->email_changeemail, $mybb->user['username'], $mybb->settings['bbname'], $mybb->user['email'], $mybb->input['email'], $mybb->settings['bburl'], $activationcode, $mybb->user['username'], $mybb->user['uid']);
+				}
+				else
+				{
+					// Email requires no activation
+					$mail_message = $lang->sprintf($lang->email_changeemail_noactivation, $mybb->user['username'], $mybb->settings['bbname'], $mybb->user['email'], $mybb->input['email'], $mybb->settings['bburl']);
+				}
+
 				$lang->emailsubject_changeemail = $lang->sprintf($lang->emailsubject_changeemail, $mybb->settings['bbname']);
-				$lang->email_changeemail = $lang->sprintf($lang->email_changeemail, $mybb->user['username'], $mybb->settings['bbname'], $mybb->user['email'], $mybb->input['email'], $mybb->settings['bburl'], $activationcode, $mybb->user['username'], $mybb->user['uid']);
-				my_mail($mybb->input['email'], $lang->emailsubject_changeemail, $lang->email_changeemail);
+				my_mail($mybb->input['email'], $lang->emailsubject_changeemail, $mail_message);
 
 				$plugins->run_hooks("usercp_do_email_verify");
 				error($lang->redirect_changeemail_activation);
